nuplan-devkit/requirements.txt
bokeh==2.4.3  # Used in the nuBoard dashboard
boto3==1.24.59  # Used in s3_store.py to download data from S3.
cachetools==5.3.0  # Used widely
casadi==3.5.5  # Used for optimization solving
control==0.9.1  # Used for LQR controller synthesis
coverage==7.2.1  # Used in computing code coverage
descartes==1.1.0  # Used in map_explorer.py
docker==6.0.1  # To run remote submissions
Fiona==1.9.1  # Used in GpkgMapsDB.py
geopandas==0.12.2  # Used to read maps
grpcio==1.43.0  # Used for communication with submission container
grpcio-tools==1.43.0  # Used for communication with submission container
guppy3==3.1.2
hydra-core==1.1.0rc1  # Used for configs
joblib==1.2.0
jupyter==1.0.0  # Used to show the notebook
jupyterlab==3.6.1
matplotlib==3.7.1  # Used for rendering
mock==5.0.1
numpy==1.22.4  # Used widely, only from this version numpy.typing.NDArray is supported
opencv-python==4.8.1.78  # Used widely
pandas==1.5.3    # Used widely
Pillow==9.4.0    # Used widely to render images
pre-commit==3.1.1  # Used in pre-commit hooks
psutil==5.9.4  # Used widely to get the number of CPUs
pyarrow==11.0.0 # For parquet
pydantic==1.10.13
pyinstrument==4.4.0    # Used widely as profiler
pyogrio==0.5.1 # A performant backend for geopandas
pyquaternion==0.9.9  # Used widely, avoid known bug with quaternion slerp
pytest==7.2.2    # Used widely
rasterio==1.3.6  # Used in GpkgMapsDB.py
ray==2.3.0 # Used widely
requests==2.28.2  # Used in BlobStore classes
retry==0.9.2
rtree==1.0.1 # Used in occupancy maps
s3fs==2023.3.0
scipy==1.10.1  # Used widely
selenium==4.8.2 # Used in bokeh export png
setuptools==59.5.0  # Used in setup.py, pinned to not break pytorch
shapely==2.0.1 # Used widely
SQLAlchemy==2.0.6  # older versions don't work with some table definitions
sympy==1.11.1  # Use for symbolic algebra
testbook==0.4.2  # Used in testing jupyter notebooks
tornado==6.2  # Used in nuboard.py
tqdm==4.65.0  # Used widely
typer==0.7.0 # Used for cli
ujson==5.7.0  # Used in serialiation_callback.py
urllib3==1.26.15  # Used in s3_store.py
